build IMUSnapshot protobuf object

+++ pkg/bno055/bno055.go
@@ -17,9 +17,9 @@ const (
 )
 
 const (
-	QuatScale        = float32(1.0 / (1 << 14))
-	EulerScale       = float32(1.0 / 16)
-	LinearAccelScale = float32(1.0 / 100)
+	QuatScale        = float64(1.0 / (1 << 14))
+	EulerScale       = float64(1.0 / 16)
+	LinearAccelScale = float64(1.0 / 100)
 )
 
 // addresses of LSBs, MSBs assumed +0x01
@@ -77,43 +77,43 @@ func (dev *Dev) readWords(startAddr byte, numWords int) ([]int16, error) {
 	return out, nil
 }
 
-func (dev *Dev) readScaledValues(startAddr byte, numWords int, scaleFactor float32) ([]float32, error) {
-	out := make([]float32, numWords)
+func (dev *Dev) readScaledValues(startAddr byte, numWords int, scaleFactor float64) ([]float64, error) {
+	out := make([]float64, numWords)
 	unscaledWords, err := dev.readWords(startAddr, numWords)
 	if err != nil {
 		return nil, err
 	}
 	for i := 0; i < numWords; i++ {
-		out[i] = float32(unscaledWords[i]) * scaleFactor
+		out[i] = float64(unscaledWords[i]) * scaleFactor
 	}
 	return out, nil
 }
 
-func (dev *Dev) ReadQuat() ([4]float32, error) {
-	var dst [4]float32
+func (dev *Dev) ReadQuat() ([4]float64, error) {
+	var dst [4]float64
 	src, err := dev.readScaledValues(QuatAddr, 4, QuatScale)
 	if err != nil {
-		return [4]float32{}, err
+		return [4]float64{}, err
 	}
 	copy(dst[:], src)
 	return dst, nil
 }
 
-func (dev *Dev) ReadEuler() ([3]float32, error) {
-	var dst [3]float32
+func (dev *Dev) ReadEuler() ([3]float64, error) {
+	var dst [3]float64
 	src, err := dev.readScaledValues(EulerAddr, 3, EulerScale)
 	if err != nil {
-		return [3]float32{}, err
+		return [3]float64{}, err
 	}
 	copy(dst[:], src)
 	return dst, nil
 }
 
-func (dev *Dev) ReadLinearAccel() ([3]float32, error) {
-	var dst [3]float32
+func (dev *Dev) ReadLinearAccel() ([3]float64, error) {
+	var dst [3]float64
 	src, err := dev.readScaledValues(LinearAccelAddr, 3, LinearAccelScale)
 	if err != nil {
-		return [3]float32{}, err
+		return [3]float64{}, err
 	}
 	copy(dst[:], src)
 	return dst, nil
